+++ 1.Utils/customSpatialCrossEntropy.m
@@ -1,6 +1,7 @@
 function loss = customSpatialCrossEntropy( Y, T )
 
 numObservations = size(Y, 4) * size(Y, 1) * size(Y, 2);
-loss_i = T .* log(customBoundAwayFromZero(single(Y)));
-loss = -sum( sum( sum( sum(loss_i, 3).*(1./numObservations), 1), 2));  
+loss_i = T .* log(customBoundAwayFromZero(double(Y)));
+loss = -sum( sum( sum( sum(double(loss_i), 3).*(1./numObservations), 1), 2));  
+
 end
